refactor(go): fold ReconnectingStream into Stream
Eliminate ReconnectingStream wrapper and StreamReceiver interface by
building reconnection directly into Stream via ConfigureReconnect().
Removes 3 generated files per SDK (reconnecting_stream.go,
reconnecting_stream_test.go, stream_receiver.go). Return type reverts
to *Stream[T] for all streaming endpoints.
