+++ generators/go-v2/sdk/src/endpoint/utils/getRawEndpointReturnTypeReference.ts
@@ -29,9 +29,7 @@ export function getRawEndpointReturnTypeReference({
         case "streaming":
             return wrapWithRawResponseType({
                 context,
-                returnType: go.Type.reference(
-                    context.getStreamReceiverTypeReference(context.getStreamPayload(body.value))
-                )
+                returnType: go.Type.reference(context.getStreamTypeReference(context.getStreamPayload(body.value)))
             });
         case "streamParameter":
             return context.getRawResponseTypeReference(go.Type.any());
